Move the pathing editor's webview script into a bundle
The .wpm editor's ~450 lines of browser JavaScript lived in a template literal
inside wpmPreview.ts, unchecked by tsc or ESLint. It is now
src/webview/wpmEditorWebview.ts, bundled to dist/webview/wpmEditorWebview.js
and loaded under a nonce; the host hands over the grid and palette data through
window.__WPM_INITIAL__ and the page no longer allows arbitrary inline scripts.
This is the pattern for the remaining inline webview scripts (model viewer,
map preview, asset browser, map-data editors, sound player).
