+++ webpack.config.js
@@ -72,6 +72,7 @@ const viewerConfig = {
 	entry: {
 		mdxViewer: './src/webview/mdxViewer.ts',
 		mpqViewerWebview: './src/webview/mpqViewerWebview.ts',
+		wpmEditorWebview: './src/webview/wpmEditorWebview.ts',
 		objModEditorWebview: './src/webview/objModEditorWebview.ts',
 	},
 	output: {
